Make searching issues/MRs by IID even easier
Prior to this change you were required to prepend your search with `#`
or `!` (like an Issue/MR link) to trigger the search by IID behaviour.
This changes things so that it's even simpler to search by IID. As long
as the input is just a number we'll ensure that the IID field is
included in search query. This also won't prevent you from searching for
the number as part of an issue title or description since we also query
those fields but we give higher weight to the IID so exact IID matches
should appear as the first result.

This MR also includes some refactoring to improve the tests.
Specifically some of the tests for searching by IID were passing by
accident before. It also refactors them to use `contain_exactly` since
this makes them more concise.