Implement new policy drawer for the policy list
This commit implements a new policy drawer for the cilium network
policies. This drawer is visible only when 'network_policy_editor' is
enabled and selected policy is a cilium policy, otherwise we are using
existing drawer implementation. Fix for the 'description' related bug
in toYaml and fromYaml helpers was also included.