Silence errors when Azure objects have aleady been deleted
In an normal upload flow, GitLab Rails will move a file to its final
destination and remove the temporary file. Workhorse issues a DELETE a
request to ensure this is cleaned up. Previously this was generating log
noise with every upload. Now only log an error message if there were
some error other than a 404.