Simple count of pre-logged-in-sessions per IP
As noted in https://gitlab.com/gitlab-org/gitlab/-/issues/219071, the
current storage mechanism is a set containing the session id, but this
is never used.
We only need the count to determine when to start potentially invoking
recaptcha.

Keeping the count only is not only more performant for the web app, it
stores significantly less in Redis, particularly for certain
pathological cases

Fixes https://gitlab.com/gitlab-org/gitlab/-/issues/219071