Allow SAST configuration with existing CI file
Previously, the Security Configuration page only linked to the SAST
Configuration page for projects _without_ an existing CI file. This was
because the initial implementation did not support updating an existing
CI file.

Now that the main technical limitation has been [lifted][1] (with
[caveats][2]), the UI can direct users to the SAST Configuration page
regardless of whether their project has an existing CI file.

This also changes the configuration button to say "Configure" if the
given feature is already configured (by any means), and "Enable" if not.
Previously this was based on whether Auto DevOps was enabled, but
checking the configured status is more direct.

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/240941.

[1]: https://gitlab.com/gitlab-org/gitlab/-/issues/232862
[2]: https://gitlab.com/gitlab-org/gitlab/-/issues/228959