Introduce snowplow event test helper
By calling the original method, we can make sure that the types
are correct. Snowplow is using a type checking gem that would
otherwise throw an exception in runtime. This is something we want
to catch in tests.

Replace the AsyncEmitter with Emitter. We want to avoid
AsyncEmitter because it does some work in a separate thread, and
we want any failures to occur in the same thread as the test.