Add warning to stop Puma and Sidekiq when restoring from backup
As we saw in https://gitlab.com/gitlab-org/gitlab/-/issues/36405, having
processes that can write to GitLab tables while restoring from a
database backup can cause duplicate rows be inserted and schema changes
to fail quietly. At the moment, we don't have a good story of restoring
a database backup in a single transaction, so we need to tell admins to
stop Puma and Sidekiq.