Add snowplow events for clicking search suggestions
This MR adds the required data-track events as described in
https://docs.gitlab.com/ee/development/telemetry/snowplow.html#tracking-in-haml-or-vue-templates
in order to allow us to measure usage of the search autocomplete
feature.

These search autocomplete links are rendered by the
deprecated_jquery_dropdown so we needed to extend this class to pass the
relevant attributes down to where the link is created.