Expose gauges to report status of partitioning
The two metrics are being reported on a per table basis. This allows
monitoring systems to detect a situation where a partition wasn't
created over an extended period of time. This is potentially a
disastrous state if a relevant partition hasn't been created yet but is
needed to create new records.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/227353