Clean up gRPC debug_error_string messages in Sentry
The upgrade of gRPC brought along a `debug_error_string` that makes it
harder for Sentry to group similar exceptions together. There doesn't
appear to be a way to turn this off, so we clean this up in a Sentry
field processsor. Now exceptions will have a field
`grpc_debug_error_string` field in the `extra` section.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/238465