Clean up gRPC debug_error_string messages in Sentry
The upgrade of gRPC brought along a `debug_error_string` that makes it
harder for Sentry to group similar exceptions together. There doesn't
appear to be a way to turn this off, so we clean this up in a Sentry
field processsor. Now exceptions will have a field
`grpc_debug_error_string` field in the `extra` section.

This was originally introduced in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/40938 but reverted
in https://gitlab.com/gitlab-org/gitlab/-/merge_requests/41083. But it's
needed because not all Gitaly calls are wrapped via `WrapsGitalyErrors`.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/238465