Update IE11 deprecation message
We have deprecated Internet Explorer 11 in 13.0 already. This removes
the extra case for it in the interface and the docs have been updated to
just indicate that we dropped it in 13.0.

The link to the feedback issue has been removed and users will still see
the outdated browser warning in case they use Internet Explorer