Strip gRPC debug_error_string from Gitaly exceptions
This reverts https://gitlab.com/gitlab-org/gitlab/-/merge_requests/40938
because it wasn't actually working when gRPC exceptions were wrapped in
`Gitlab::Git::CommandError` via `wrapped_gitaly_errors`.

A much simpler approach is to strip `debug_error_string` when a
`Gitlab::Git::BaseError` is created. With this change, we have
effectively the same exception strings as we did before the upgrade of
gRPC.

Note that it appears Sentry retains a backtrace of the original gRPC
exception that contains the `debug_error_string`, so if we really want
to look at it the information is still there.

This should solve https://gitlab.com/gitlab-org/gitlab/-/issues/238465.