Limit the context for paused elasticsearch jobs
When we pause elasticsearch indexing, we store the jobs in a ZSET. The
advantage of that would be that the jobs are automatically
deduplicated. Thanks Redis!

However, since the application context contains unique
information (correlation_id, user, caller_id), we would lose this
benefit.

With this, we only include the shared information (the `meta.project`
key), which is shared across the deduplicatable jobs.

The project information should be enough context for logging purposes:
The namespace can be inferred from that, and the caller_id should be
populated again when the job is rescheduled.