Change 2FA to verify password hash instead of timestamp
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/41327, we fixed
a problem that prevented LDAP users from logging in due to an issue with
timestamp precision. To avoid pitfalls with time comparisons, compare a
hash of the encrypted password instead of the last updated timestamps.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/244638.