Fix formatting of restore from backup text
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/40911, we added
a confirmation step if psql encountered any errors. However, the threads
used to read stdout/stderr were not joined, so it was likely that the
stdout spew from psql would continue to spew before the process was
actually complete. As done in Popen3#capture3, this change waits for
those stdout/stderr threads to finish before proceeding.