Add incident label for manually created incident issues
Previously, we've created and applied the "incident" label
only when calling the `IncidentManagment::Incidents::CreateService`
explicitly. After we've added the issue type "incident" incidents
can be created/updated using the `Issues::Create/UpdateService`.
Those services did not apply this label.

This commit moves the code which applies the label close to the
`Issue` services to make sure we are labeling all incidents with
the "incident" label correctly.