Add DevOps Score header in documentation
Recently the DevOps Score feature was moved to a new page called
DevOps Reports. At the same time, the documentation page for the feature
was moved to a new page for DevOps Reports. This adds a section header
to the documentation page for "DevOps Score" to make it clear that this
feature appears on this page. This will help with searchability as well.