Remove stale merge refs
14 days after a merge request gets closed/merged, its merge ref
will be deleted and moved as a keep-around ref (if not yet added).

This is to ensure that the commit won't be deleted when GC runs
and still being able to reduce the number of advertised refs on
git push/pull.

Before deletion, it'll be cached to `merge_ref_sha` column of
`merge_requests` table so we can still display the merge ref diff.