Download & parse only the necessary CI job artifact
Previously we were downloading & parsing all CI job artifacts for each
request to fetch the list of findings of a pipeline. This was causing
timeout errors for pipelines with tens of security scans.

With this change, we are downloading & parsing only the necessary CI
job artifacts which should perform better than the previous version.