Refuse to perform an LFS clean on projects that are fork roots
When a project is at the root of a fork network, an intersection of
technical debt means that all child forks have a claim on the LFS
objects owned by the fork parent. This makes it dangerous to remove
any LFS references in the fork root without checking all the repos
for all the children, which quickly mounts up so as to be unmanageable.

For now, just skip projects like this.

If a fork is an internal or leaf node in the fork graph, they may have
some LFS objects of their own. Those references are safe to prune.