Make Pages disk access fail in web server
Previously, when `GITLAB_PAGES_DENY_DISK_ACCESS` was set, accessing
`Gitlab::Pages::Settings.path` from a web process would track an
exception and continue on.

Now it will simply raise the exception and fail. This is to make it
easier for developers to find a problem with their code before taking it
to production on GitLab.com, where the Pages NFS mount will soon be
removed.