Atomically create table and its partitions
This is to avoid a race condition when we succeed to create the
partitioned table but failed to create its partitions.

In that case, automatic partition management might kick in leading to
undesired states.

We basically want the migration to fully succeed first, before
partitions are being automatically managed.

Relates to https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/5597