Fix merge request chat messages for adding and removing approvals
With GitLab 13.2 we have ported optional MR approvals to Core. If a user
uses an integration like Slack or Mattermost, the chat message will not
send the correct chat message. We can simply backport the methods to CE
in order to fix this problem.

See https://gitlab.com/gitlab-org/gitlab/-/issues/28138