We only need to test the value of `scoped_items`
The changes tested here are to see if in EE projects in the same groups
count as part of the same relative positioning set. We do not need to
test every method - just the scoping query.

Frankly a bunch of the movement tests could also be removed, but they
are all calling public methods so it is fine for them to test the
interface.