Log authentication events alongside existing audit events
For authentication, add both an audit event and authentication
event for the short term. Eventually we will diverge and only
store authentication events when an authentication occours.
At that time, authentication events will be a subclass of audit
events. For now we store both and use authentication events for
usage ping exclusively.