Add `_name` to ES queries to improve readability
This MR adds a new helper class: `Elastic::Latest::QueryFactory` that
can be used to generate semantic names for queries.

In this MR, a hierarchical pattern has been used, to correlate
different part of the queries, such as:

- `issue:match:search_query`
- `issue:related:project`
- `doc:is_a:blob`

This MR sets a lot of names across the query parts, so it is expected
that some of these names might be sub-optimal.