Find admin integration with group-level integrations
After adding the group_id column to the services table
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/38499, we can now
save integrations that belongs to a group
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/39959.

Integrations can inherit the admin integration from the UI selector.
Now, having group-level integration we need to find the right
integration to inherit the values from.

We have to find the closest group integration or instance-level
integration.