Fix hover style in dark mode
Our page specific CSS don't have a dark mode variant. Rather than
compiling a page specific CSS file for each theme, we make use of CSS
variables in order to define the correct colors. There is prior art with
the IDE Themes and eventually we can move to _only_ use these variables.

In case one variable is undefined, we will fallback on the light theme
value for it. Thus we just need to define the variables in the dark
theme, since the fallback _is_ the light theme.

This approach will allow us to create other themes, e.g. for
accessibility (color blindness, high contrast) in the future.