Convert group wikis into a licensed feature
This was a misunderstanding on my part since licensed features default
to on, but we have a `beta_feature_available?` helper specifically to
handle licensed features which should default to off.

This also reverts the previous refactoring to the `feature_available?`
helper in `GroupPolicy`, which is now redundant and removed.