Exclude tmp dirs from backups
Add '--exclude' for tmp directories in located in paths included in
backups.  Currently a backup job may fail due to a file or directory
within a tmp dir being removed during the backup.

This change adds a new 'exclude' optional parameter to Backup::Files
to indicate directories that should not be included in backups. This
should prevent changes in temporary files from breaking backups and
also stop unneeded files from being included in the backup archive.