Track projects using code intelligence
We need to track the usage of code intelligence.

The authorize endpoint is called by workhorse to determine if it
needs to parse LSIF which is needed for code intelligence feature
to work.

This adds tracking of `i_source_code_code_intelligence` event
whenever it's determined that we are parsing LSIF. Uses the HLL
counter so we get distinct count of projects.