Remove job logs from notification e-mails
Previously we included the last 30 lines of a job log in notification
e-mails to help give some details as to why a pipeline failed, but
including these logs in the e-mails has several issues:

1. Often more than 30 lines are needed to understand a build failure.
2. These logs could leak sensitive information.
3. The notification e-mails are scheduled around the same time when the
job log is archived to object storage, so the notification service may
need to read from a shared volume (NFS). There are race conditions when
the file has been archived, and we have to retry when these happen.

To avoid these pitfalls, we omit the logs from notification e-mails.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/247117