Make startup.js to account for the param orders
When startup.js checks whether a API call has been prefetched or not,
it uses a path string (e.g., /path?a=1&c=2) but this check can fail
when the params listed in the path string in a different order.

This MR uses options.sort for mergeUrlParams
to sort the params in the requested path string to fix the above issue.