Fix reindexing bug
When a `statement_timeout` is set, this leads to trouble with the
inclusion of the migration helper.

There is basically a name clash between the reindexing class with
`#execute` and the migration helper for the statement timeout expecting
to be able to execute SQL queries through `#execute`.

This has been fixed with renaming the method on the reindexing class.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/251084