Store manifest import data outside of session
When a user uploads a manifest file to import (typically for the Android
Open Source Project), we store some data about that import in Redis.
Previously, we stored that in the session directly, but that meant that
as long as the session didn't expire, neither would this - quite large -
import status data.

This changes the import status data to be stored in its own dedicated
fields, with explicit expiry set. To handle existing imports, it will
still read data from the session if it can't find anything in the new
place.