Make concern module more auto-load friendly
The existing folder `app/controllers/admin/concerns` is not part of the
top-level auto-load and making EE extension for this concern is
problematic. Other non-admin controller
(e.g. `OmniauthCallbacksController`) is also using this concern so it is
better place in the standard `app/controllers/concerns` folder.

The class is also renamed to be consistent with the existing counterpart
`AuthenticatesWithTwoFactor` class and avoid issue with RSpec cop
misreading the name due to usages of numbers and acronyms (i.e. `2FA`).