Update programming language colors and metadata
When linguist was phased out the data we depended on was vendored. This
created some stale data since than, and this commit updates the vendored
file to the file upstream has at 7.11.0. This version number was choosen
as Gitaly will shortly run `github-linguist` at that version.

Than this updated file is used to update the colors in the database, for
records already in the database. About 200 `UPDATE` statements are
executed.