Fix errors when checking MR task lists
The HTML spec in https://www.w3.org/TR/html52/syntax.html#in-body
mentions: Newlines at the start of textarea elements are ignored as
an authoring convenience.

This causes issues with task list parsing because the lines are now off
by one.

We didn't have this problem with issues because the textarea value was
set in Javascript with Vue.

This also removes instances of js-task-list-field which are actually
unused.