Remove switch to basic search link
This was originally implemented as a back up because we were concerned
about the reliability of Elasticsearch. We assumed it would be safe to
add this link just in case our Elasticsearch experience was consistently
bad.

But it turned out that we forgot to ever turn the feature flag on and
our Elasticsearch has been quite robust anyway. As such we may as well
remove the feature now.

We still keep the hidden feature where the user can manually add a
`basic_search=true` to the URL to get the basic search functionality
because this is still useful for diagnostic purposes but we don't need
to provide a link to this as it's not something most people will need to
use.