Permanently enable package_file_registries field
The feature flag `:geo_package_file_replication` is enabled by default
because we have released the feature. The feature flag check remains, so
that *replication* can be quickly and easily disabled in case a customer
does not need to replicate this resource, or in case there is a serious
bug or infrastructure problem causing excessive replication errors.

At the moment, there is no way to specify that the GraphQL field's
feature flag should default enabled. Additionally, there is no need to
disable the GraphQL API field, regardless of this feature flag, now that
the feature is released. It just exposes the `package_file_registry`
table.

Therefore, this commit safely resolves a bug in which the
`Admin > Geo > Replication > Package files` view (which is enabled by
default because `:geo_package_file_replication` is enabled by default)
cannot access the GraphQL field unless the feature is *explicitly and
manually* enabled.

See https://gitlab.com/gitlab-org/gitlab/-/issues/255280