Refactor ComplianceFramework in to AR Model
Builds on this MR
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/43301

To use the new table to represent
compliance frameworks.

Adds foreign key constraint and
index between
ComplianceManagement::ComplianceFramework::ProjectSettings
and ComplianceManagement::Framework

Updates the use of the `framework` attribute to
use the new foreign key relation. This commit is
large due to the number of places that it is
referenced throughout the codebase.