Create BulkImport::Configuration model for auth
We want to start importing Group and Project data directly from
another GitLab instance, via the API. To do this, we need somewhere
to store authentication data so that we can execute the import in
background jobs.

Create the BulkImport::Configuration model which is associated with
a BulkImport, and used to store the authentication data for the
source GitLab instance. This will be followed up by the
introduction of the other models the store the import data.

This model introduces:

1. `bulk_import_id` -> the bulk import that this configuration
   applies to.
2. `api_url` -> the domain for the source instance.
3. `access_token` -> the personal access token of the user on the
   source instance for authentication.

This is a component of the Group Migration MVC epic:
https://gitlab.com/groups/gitlab-org/-/epics/4374

MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/42978
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/250280