Fix avatar crop screen in dark mode
Use a regular scss import for cropper as otherwise it was
broken in dark mode. Now works the same as select2. There
is a slight change in order but that is probably fine

Also remove the require_self call since it no longer does
anything as there are no other requires here