Extract CSS utilities into separate files
When we extract page specific CSS, they sometimes overwrite our utility
classes and and force us to use important! utility classes. [0]

In order to make our utilities take precendence, we split them off, so
we can import them later. Note that we have to do it once for light mode
and dark mode.

[0]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/42795#note_415620902