Fix Webpack manifest not working with single-threaded Web servers
The removal of the webpack-rails gem in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/42485 changed the
way the manifest was loaded from the Webpack dev server. Instead of
contacting directly to the HTTP port 3809, the change caused an external
HTTP(s) request to be made to the GDK URL
(e.g. https://gdk.test:3001/assets/manifest.json).  This would cause a
single-threaded like Thin to lock up.

This change restores the separate `manifest_host` and `manifest_port`
that was previously used to contact the Webpack server directly.