Replace dynamic QA selectors with generic hard-coded values
The existing QA selectors were in-place due to the process of migrating
the audit events from HAML templates to a generic Vue app. This removes
these selectors and replaces them with hard-coded generic values as the
dynamic values are no longer needed. We can use the generic values very
simply for all the QA testing requirements.

We can also remove the audit_log_filter QA selector as it isn't used