Add the GlModal to SSH keys deletion
Updated the admin and user SSH key delete confirmation to use GlModal.
This uses the shared modal feature to trigger the GlModal on click
of the delete buttons and processes the deletion on confirmation.

Also revised the wording of the buttons to "Delete" from "Remove"
to make sure that the wording is consistent across the product.

Updated the specs to work with the new modal as well