Add Group.mergeRequests field
This allows one to request the merge requests for a particular group,
including all of its projects and subgroups.

This functionality is not available in any other way at present.

To achieve this, a new resolver is added: GroupMergeRequestResolver,
which is tested with request tests.

The following minor additions are made:

- The merge requests factory now is able to build merge requests with
  unique authors (this is helpful to distinguish querying by project and
  by author, since otherwise they are the same).

- Abstract out include_subgroups, author and assignee filters
  this moves some filter definitions to specialised DSL
  methods, so they can be used where it makes sense without repeating the
  field definitions.