Fix missing active constant error
When we refer to a constant from initializers which is supposed to be
loaded by `ActiveSupport::Dependencies`, it's possible to get an error
in the development environment if Rails tries to load/reload a missing
constant under the namespace of that constant.

The reason is since the constant is loaded in initializers, it stays in
the memory and conflicts with the one reloaded by Rails.

This commit fixes only the `Gitlab::Metrics::Subscribers::ActiveRecord`
case which has been used in lograge configuration.