Fixes two button submitting of merge request reviews
Improves the experience of submitting reviews
in merge requests by removing the need to open the dropdown
and then submit the review.
Instead the user can now submit the review by clicking the
submit review button.

This also removes the discard review modal as each review
note has an option to be deleted individually.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/212334