Check for default_max_patch_size changes
If the value is updated by an admin, we may _want_ to display the
highlighted lines on a page, but will be unable to deliver them as the
cached version might have been set as empty ([]) when the file
PREVIOUSLY was considered "too large". When one of these situations is
detected - the file is unreadable from the cache, and would've violated
the default value but NOT violate the current value, go ahead and
manually update the cache with it.