Use local QueryContext query context instance
This commit changes the way the `QueryContext` (renamed from
`QueryFactory`) is used in the query building process.

With this commit, a `context` method is defined in the root
`ApplicationClassProxy` and used through the whole process.

This simplifies the logic quite a bit and removes the need for a thread
locality.