Add filters to the milestones resolver
This adds title filters to the milestones resolver, and a simplified
form of the date search, which makes it easier to search for milestones
overlapping a given date.

Tests are added for `within_timeframe`. (this involves subtle
differences for Milestone and Iteration, since they have different
validations)

To improve types and error handling for clients, the separate
startDate/endDate arguments are deprecated in favor of a custom time
range type.