Cache markdown when viewing a blob
In order to improve performance of
`Projects::BlobController#show.json` when viewing a markdown blob,
we need to utilize the cache so succeeding requests to the same
blob won't have to re-parse markdown.

Based on tests done locally, this improves the response time of
the endpoint when viewing a large markdown file from ~11s to ~4s.