This adds a status label to http_requests_total
This adds a status label to the http_requests_total metric. To do so,
we've had to move the recording to an ensure block after the request
is performed.

In practise our @app.call never fails. But there's a spec covering the
recording of the request count even if it fails. That's why the
recording happens inside an ensure.