Add range and date types
The range type abstracts over pairs of ordered start and end of the same
type. The date type is a better version of the one in the GraphQL
library, since it raises coercion errors rather than returning nil.

Timeframe is implemented as `Range[Date]`, which is a much better
description of what it is.