Update to Rack v2.1.4
This is needed to address a few outstanding CVEs and fix cookie
timestamp formats.

Full list of changes:
https://github.com/rack/rack/compare/2.0.9..2.1.4

Relates to:
* https://gitlab.com/gitlab-org/gitlab/-/issues/36362
* https://gitlab.com/gitlab-org/gitlab/-/issues/228622

Rack v2.1.x no longer coerces the body to a string. Previously in a few
places the Grape API was returning the status code as an integer, which
Grape used as the response body. To preserve the legacy behavior, we
explicitly set the body to the stringified integer.