Whats New - Add "infinite" scroll/pagination
Adding infinite scroll and "pagination" to the "whats new" component.

Something worth mentioning is the integration between the drawer and the
infinite scroll. The Drawer component has the overflow-y attribute
declared and it conflicts with the infinite scroll component
receiving a scroll event. My solution is to hide any overflow from
the drawer, and set a fixed pixel amount to the infinite scroll
as a prop.

part of:
https://gitlab.com/gitlab-org/growth/engineering/-/issues/5388