Memoize repository storage shard in GroupWiki
This caused a bug where sometimes one shard was used for the repository
creation in `Wiki#create_wiki_repository`, but then a different one was
persisted to the DB in `GroupWiki#create_wiki_repository`.

So subsequent requests would use the wrong shard, and run into
`Gitlab::Git::Repository::NoRepository` errors.