Only generate CI artifact metadata for ZIP files
In the beginning, CI artifacts were resticted to ZIP files, but they
have been expanded to be JSON files, tar.gz files, and more. Prior to
this change, these gitlab-zip-metadata would be run for each of these
non-ZIP artifacts, generating lots of noise in the logs.

We now inspect the `artifact_format` query string to determine whether
to generate metadata. If the format is ZIP, which is the case for CI
archives and LSIF files, we generate metadata.

Relates to https://gitlab.com/gitlab-org/gitlab-workhorse/-/issues/272